Expression: do not clamp large integer literals to INT_MAX (#23419)
essentiallyInteger() converted a positive value larger than INT_MAX but
within long range with static_cast<int> instead of static_cast<long>, so
integer literals such as 9999999999 were clamped to 2147483647 (the
negative branch was already correct). Results on int and float
properties and evalExpression() were all affected.

Cast to long to match the return code (2 = value fits in a long).
Extends the dimensionless-integer parser test.

Fixes #23419

+++ src/App/Expression.cpp
@@ -350,7 +350,7 @@ static inline int essentiallyInteger(double a, long &l, int &i) {
             return 1;
         }
         else if (intpart <= static_cast<double>(std::numeric_limits<long>::max())) {
-            l = static_cast<int>(intpart);
+            l = static_cast<long>(intpart);
             return 2;
         }
     }
